requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-appconf==1.0.2
django-compat==1.0.14
dj-database-url==0.4.2
django-debug-toolbar==1.8
django-extensions==1.9.0
django-filter==1.0.4
django-hijack==2.1.4
django-hijack-admin==2.1.4
django-user-accounts==2.0.3
djangorestframework==3.6.4
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pytz==2017.2
requests==2.18.4
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.22
wheel==0.24.0
whitenoise==3.3.0
psycopg2==2.6.2